+++ lib/argument.js
@@ -75,4 +75,21 @@ class Argument {
   };
 }
 
+/**
+ * Takes an argument and returns its human readable equivalent for help usage.
+ *
+ * @param {Argument} arg
+ * @return {string}
+ * @api private
+ */
+
+function humanReadableArgName(arg) {
+  const nameOutput = arg.name() + (arg.variadic === true ? '...' : '');
+
+  return arg.required
+    ? '<' + nameOutput + '>'
+    : '[' + nameOutput + ']';
+}
+
 exports.Argument = Argument;
+exports.humanReadableArgName = humanReadableArgName;
